refactor: move input utility to inputs folder
